Rearrange the files in SimG4CMS/Forward and collaspse some of the .h and .cc files
