+++ SimG4CMS/Forward/plugins/BscTest.cc
@@ -2,23 +2,37 @@
 //
 
 // system include files
+#include <cmath>
 #include <iostream>
 #include <iomanip>
-#include <cmath>
+#include <map>
+#include <string>
 #include <vector>
-//
-#include "SimG4Core/Notification/interface/BeginOfEvent.h"
-#include "SimG4Core/Notification/interface/EndOfEvent.h"
-#include "SimG4Core/Notification/interface/TrackWithHistory.h"
-#include "SimG4Core/Notification/interface/TrackInformation.h"
 
+//
 #include "FWCore/Framework/interface/Event.h"
 #include "FWCore/Framework/interface/EventSetup.h"
+#include "FWCore/Framework/interface/Frameworkfwd.h"
+#include "FWCore/Framework/interface/one/EDAnalyzer.h"
 #include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+
+#include "SimG4Core/Notification/interface/TrackWithHistory.h"
+#include "SimG4Core/Notification/interface/TrackInformation.h"
+#include "SimG4Core/Notification/interface/Observer.h"
+#include "SimG4Core/Notification/interface/BeginOfJob.h"
+#include "SimG4Core/Notification/interface/BeginOfRun.h"
+#include "SimG4Core/Notification/interface/EndOfRun.h"
+#include "SimG4Core/Notification/interface/BeginOfEvent.h"
+#include "SimG4Core/Notification/interface/EndOfEvent.h"
+#include "SimG4Core/Notification/interface/BeginOfTrack.h"
+#include "SimG4Core/Notification/interface/EndOfTrack.h"
+#include "SimG4Core/Watcher/interface/SimWatcher.h"
+
 // to retreive hits
+#include "SimG4CMS/Forward/interface/BscG4Hit.h"
 #include "SimG4CMS/Forward/interface/BscNumberingScheme.h"
 #include "SimG4CMS/Forward/interface/BscG4HitCollection.h"
-#include "SimG4CMS/Forward/interface/BscTest.h"
 
 // G4 stuff
 #include "G4SDManager.hh"
@@ -29,20 +43,190 @@
 #include "G4UserEventAction.hh"
 #include "G4TransportationManager.hh"
 #include "G4ProcessManager.hh"
+#include "G4VTouchable.hh"
 
+#include "CLHEP/Vector/ThreeVector.h"
+#include "CLHEP/Vector/LorentzVector.h"
+#include "CLHEP/Random/Randomize.h"
 #include "CLHEP/Units/GlobalSystemOfUnits.h"
 #include "CLHEP/Units/GlobalPhysicalConstants.h"
 
-//================================================================
-// Root stuff
-
-// Include the standard header <cassert> to effectively include
-// the standard header <assert.h> within the std namespace.
-#include <cassert>
-
-//================================================================
+// ----------------------------------------------------------------
+// Includes needed for Root ntupling
+//
+#include "TROOT.h"
+#include "TFile.h"
+#include "TH1.h"
+#include "TH2.h"
+#include "TProfile.h"
+#include "TNtuple.h"
+#include "TRandom.h"
+#include "TLorentzVector.h"
+#include "TUnixSystem.h"
+#include "TSystem.h"
+#include "TMath.h"
+#include "TF1.h"
+#include "TObjArray.h"
+#include "TObjString.h"
+#include "TNamed.h"
 
 //#define EDM_ML_DEBUG
+
+class BscAnalysisHistManager : public TNamed {
+public:
+  BscAnalysisHistManager(const TString& managername);
+  ~BscAnalysisHistManager() override;
+
+  TH1F* GetHisto(Int_t Number);
+  TH1F* GetHisto(const TObjString& histname);
+
+  TH2F* GetHisto2(Int_t Number);
+  TH2F* GetHisto2(const TObjString& histname);
+
+  void WriteToFile(const TString& fOutputFile, const TString& fRecreateFile);
+
+private:
+  void BookHistos();
+  void StoreWeights();
+  void HistInit(const char* name, const char* title, Int_t nbinsx, Axis_t xlow, Axis_t xup);
+  void HistInit(
+      const char* name, const char* title, Int_t nbinsx, Axis_t xlow, Axis_t xup, Int_t nbinsy, Axis_t ylow, Axis_t yup);
+
+  const char* fTypeTitle;
+  TObjArray* fHistArray;
+  TObjArray* fHistNamesArray;
+};
+
+
+/*
+class ObserveBeginOfRun : private Observer<const BeginOfRun *> 
+{
+public:
+    ObserveBeginOfRun();
+private:
+    void update(const BeginOfRun * run);
+};
+
+class ObserveEndOfRun : private Observer<const EndOfRun *> 
+{
+public:
+    ObserveEndOfRun();
+private:
+    void update(const EndOfRun * run);
+};
+
+class ObserveBeginOfEvent : private Observer<const BeginOfEvent *> 
+{
+public:
+    ObserveBeginOfEvent();
+private:
+    void update(const BeginOfEvent * evt);
+};
+
+class ObserveEndOfEvent : private Observer<const EndOfEvent *> 
+{
+public:
+    ObserveEndOfEvent();
+private:
+    void update(const EndOfEvent * evt);
+    //    std::vector<BscG4Hit> theStripHits;
+};
+
+class ObserveBeginOfTrack : private Observer<const BeginOfTrack *> 
+{
+public:
+    ObserveBeginOfTrack();
+private:
+    void update(const BeginOfTrack * trk);
+};
+
+class ObserveEndOfTrack : private Observer<const EndOfTrack *> 
+{
+public:
+    ObserveEndOfTrack();
+private:
+    void update(const EndOfTrack * trk);
+};
+
+class ObserveStep : private Observer<const G4Step *> 
+{
+public:
+    ObserveStep();
+private:
+    void update(const G4Step * step);
+};
+*/
+
+class BscTest : public SimWatcher,
+                public Observer<const BeginOfJob*>,
+                public Observer<const BeginOfRun*>,
+                public Observer<const EndOfRun*>,
+                public Observer<const BeginOfEvent*>,
+                public Observer<const BeginOfTrack*>,
+                public Observer<const G4Step*>,
+                public Observer<const EndOfTrack*>,
+                public Observer<const EndOfEvent*> {
+public:
+  BscTest(const edm::ParameterSet& p);
+  ~BscTest() override;
+  //MyActions();
+  //MyActions();
+private:
+  // observer classes
+  void update(const BeginOfJob* run) override;
+  void update(const BeginOfRun* run) override;
+  void update(const EndOfRun* run) override;
+  void update(const BeginOfEvent* evt) override;
+  void update(const BeginOfTrack* trk) override;
+  void update(const G4Step* step) override;
+  void update(const EndOfTrack* trk) override;
+  void update(const EndOfEvent* evt) override;
+
+private:
+  //UHB_Analysis* UserNtuples;
+  BscNumberingScheme* theBscNumberingScheme;
+
+  int iev;
+  int itrk;
+  G4double entot0, tracklength0;
+
+private:
+  // Utilities to get detector levels during a step
+
+  int detLevels(const G4VTouchable*) const;
+  G4String detName(const G4VTouchable*, int, int) const;
+  void detectorLevel(const G4VTouchable*, int&, int*, G4String*) const;
+
+  double rinCalo, zinCalo;
+  int lastTrackID;
+  int verbosity;
+
+  // SumEnerDeposit - all deposited energy on all steps ;  SumStepl - length in steel !!!
+  G4double SumEnerDeposit, SumStepl, SumStepc;
+  // numofpart - # particles produced along primary track
+  int numofpart;
+  // last point of the primary track
+  G4ThreeVector lastpo;
+
+  // z:
+  double z1, z2, z3, z4;
+
+private:
+  Float_t bsceventarray[1];
+  TNtuple* bsceventntuple;
+  TFile bscOutputFile;
+  int whichevent;
+
+  BscAnalysisHistManager* TheHistManager;  //Histogram Manager of the analysis
+  std::string fDataLabel;                  // Data type label
+  std::string fOutputFile;                 // The output file name
+  std::string fRecreateFile;               // Recreate the file flag, default="RECREATE"
+
+  //  //  //  //  //  //  TObjString fHistType;
+  //   TString fDataLabel;             // Data type label
+  //   TString fOutputFile;            // The output file name
+  //   TString fRecreateFile;          // Recreate the file flag, default="RECREATE"
+};
 
 //UserVerbosity BscTest::std::cout("BscTest","info","BscTest");
 enum ntbsc_elements { ntbsc_evt };
@@ -845,3 +1029,8 @@ void BscTest::update(const EndOfEvent* evt) {
     edm::LogVerbatim("BscTest") << "BscTest:  END OF Event " << (*evt)()->GetEventID();
   }
 }
+
+#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+
+DEFINE_SIMWATCHER(BscTest);
