+++ SimG4CMS/Forward/plugins/DoCastorAnalysis.cc
@@ -9,22 +9,102 @@
 // Original Author: P. Katsas
 //         Created: 02/2007
 //
-#include "FWCore/MessageLogger/interface/MessageLogger.h"
-#include "SimG4CMS/Calo/interface/CaloG4Hit.h"
-#include "SimG4CMS/Calo/interface/CaloG4HitCollection.h"
+
 #include "DataFormats/Math/interface/Point3D.h"
 
-#include "SimG4CMS/Forward/interface/DoCastorAnalysis.h"
+#include "FWCore/MessageLogger/interface/MessageLogger.h"
+#include "FWCore/ParameterSet/interface/ParameterSet.h"
+
+#include "SimG4CMS/Calo/interface/CaloG4Hit.h"
+#include "SimG4CMS/Calo/interface/CaloG4HitCollection.h"
 #include "SimG4CMS/Forward/interface/CastorNumberingScheme.h"
 
+#include "SimG4Core/Notification/interface/BeginOfJob.h"
+#include "SimG4Core/Notification/interface/BeginOfRun.h"
+#include "SimG4Core/Notification/interface/EndOfRun.h"
+#include "SimG4Core/Notification/interface/BeginOfEvent.h"
+#include "SimG4Core/Notification/interface/EndOfEvent.h"
+#include "SimG4Core/Notification/interface/Observer.h"
+#include "SimG4Core/Watcher/interface/SimWatcher.h"
+
+#include "G4SDManager.hh"
+#include "G4Step.hh"
+#include "G4Track.hh"
+#include "G4Event.hh"
+#include "G4PrimaryVertex.hh"
+#include "G4VProcess.hh"
+#include "G4HCofThisEvent.hh"
+#include "G4UserEventAction.hh"
+#include "CLHEP/Units/GlobalSystemOfUnits.h"
+#include "CLHEP/Units/GlobalPhysicalConstants.h"
+#include "CLHEP/Random/Randomize.h"
+
+#include "TROOT.h"
 #include "TFile.h"
+#include "TH1.h"
+#include "TH2.h"
+#include "TProfile.h"
+#include "TNtuple.h"
+#include "TRandom.h"
+#include "TLorentzVector.h"
+#include "TUnixSystem.h"
+#include "TSystem.h"
+#include "TMath.h"
+#include "TF1.h"
+
+#include <cassert>
 #include <cmath>
 #include <iostream>
 #include <iomanip>
+#include <map>
 #include <memory>
+#include <string>
 #include <vector>
 
 //#define EDM_ML_DEBUG
+
+class DoCastorAnalysis : public SimWatcher,
+                         public Observer<const BeginOfJob *>,
+                         public Observer<const BeginOfRun *>,
+                         public Observer<const EndOfRun *>,
+                         public Observer<const BeginOfEvent *>,
+                         public Observer<const EndOfEvent *>,
+                         public Observer<const G4Step *> {
+public:
+  DoCastorAnalysis(const edm::ParameterSet &p);
+  ~DoCastorAnalysis() override;
+
+private:
+  // observer classes
+  void update(const BeginOfJob *run) override;
+  void update(const BeginOfRun *run) override;
+  void update(const EndOfRun *run) override;
+  void update(const BeginOfEvent *evt) override;
+  void update(const EndOfEvent *evt) override;
+  void update(const G4Step *step) override;
+
+private:
+  int verbosity;
+
+  std::string TreeFileName;
+
+  TFile *CastorOutputEventFile;
+  TTree *CastorTree;
+
+  int eventIndex;
+
+  std::vector<double> simhit_x, simhit_y, simhit_z;
+  std::vector<double> simhit_eta, simhit_phi, simhit_energy;
+  std::vector<int> simhit_sector, simhit_module;
+  //std::vector<double> simhit_time;
+
+  std::vector<double> *psimhit_x, *psimhit_y, *psimhit_z;
+  std::vector<double> *psimhit_eta, *psimhit_phi, *psimhit_energy;
+  std::vector<int> *psimhit_sector, *psimhit_module;
+  //std::vector<double> *psimhit_time;
+
+  double simhit_etot;
+};
 
 DoCastorAnalysis::DoCastorAnalysis(const edm::ParameterSet& p) {
   edm::ParameterSet m_Anal = p.getParameter<edm::ParameterSet>("DoCastorAnalysis");
@@ -224,3 +304,8 @@ void DoCastorAnalysis::update(const EndOfEvent* evt) {
 void DoCastorAnalysis::update(const EndOfRun* run) { ; }
 
 void DoCastorAnalysis::update(const G4Step* aStep) { ; }
+
+#include "SimG4Core/Watcher/interface/SimWatcherFactory.h"
+#include "FWCore/PluginManager/interface/ModuleDef.h"
+
+DEFINE_SIMWATCHER(DoCastorAnalysis);
