+++ SimG4CMS/Forward/src/ZdcSD.cc
@@ -243,15 +243,17 @@ double ZdcSD::getEnergyDeposit(const G4Step* aStep) {
           float tan_arcos = 2. * a * d;
           if (tan_arcos != 0.)
             arg_arcos = (r * r - a * a - d * d) / tan_arcos;
-          // std::cout.testOut << "  d_qz: " << r << "," << a << "," << d << " " << tan_arcos << " " << arg_arcos;
           arg_arcos = std::abs(arg_arcos);
-          // std::cout.testOut << "," << arg_arcos;
           float th_arcos = acos(std::min(std::max(arg_arcos, -1.f), 1.f));
-          // std::cout.testOut << " " << th_arcos;
           d_qz = th_arcos / twopi;
-          // std::cout.testOut << " " << d_qz;
           d_qz = std::abs(d_qz);
-          // std::cout.testOut << "," << d_qz;
+#ifdef EDM_ML_DEBUG
+          edm::LogVerbatim("ForwardSim").testOut << "  d_qz: " << r << "," << a << "," << d << " " << tan_arcos << " " << arg_arcos;
+          edm::LogVerbatim("ForwardSim").testOut << "," << arg_arcos;
+          edm::LogVerbatim("ForwardSim").testOut << " " << d_qz;
+          edm::LogVerbatim("ForwardSim").testOut << " " << th_arcos;
+          edm::LogVerbatim("ForwardSim").testOut << "," << d_qz;
+#endif
         }
       }
     }
